Rename vscode.ts to vscodeutils.ts
